refactor: rename to utils finally
