+++ pkg/utils/tls_test.go
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (C) 2023 Intel Corporation
 
-// Package server implements the server
-package server
+// Package utils contails useful helper functions
+package utils
 
 import (
 	"crypto/tls"
